Use observables with duplex communication
The app uses redux as a state management utility, which means the
communication inside the app will be done using actions `{ type: String,
...body }`. Because actions are pure data passing them around using
observables is a way to model the data flow inside the code.

This PR refactors what was shown in the precedent PR (see first commit)
to enable a kind of "duplex" communication. The `createSocket$` utility
now takes an `action$` stream as input. Upon subscription it will create
a WS instance and connect (could be parameterized and made lazy easily).

It will then look for several actions with "special" types (qualified,
could be parameterized also). For instance if an action with type `
"SOCKET_SEND_MESSAGE"` is dispatched upstream, then if will convert
into JSON the action and send it via the websocket. When a message is
received a `"SOCKET_RECEIVED_MESSAGE"` is dispatched to consumers.

This means that the behaviour of the WS can be controlled entirely
through actions.